requirementsBackup.txt
aioesphomeapi==21.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.3.2
aiohttp==3.9.3
aiosignal==1.3.1
ajsonrpc==1.2.0
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.3
argcomplete==3.2.2
argilla==1.29.0
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
Authlib==1.2.0
av==10.0.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bidict==0.22.1
bitarray==2.9.2
bitstring==4.1.4
black==23.3.0
blinker==1.6.2
bottle==0.12.25
bs4==0.0.1
cachetools==5.3.2
certifi==2023.5.7
cffi==1.15.1
chacha20poly1305-reuseable==0.12.1
chardet==5.1.0
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.7
cohere==4.44
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.0
commonmark==0.9.1
configobj==5.0.8
configparser==7.0.0
contourpy==1.1.1
cryptography==42.0.2
ctranslate2==3.16.1
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.14.5
decorator==5.1.1
deepgram-sdk==3.2.2
Deprecated==1.2.14
dill==0.3.7
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
EbookLib==0.18
ecdsa==0.18.0
effdet==0.4.1
elevenlabs==0.2.27
elevenlabslib==0.10.0
emoji==2.8.0
esphome==2023.12.9
esphome-dashboard==20231107.0
esptool==4.6.2
et-xmlfile==1.1.0
etelemetry==0.3.1
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.104.0
fastavro==1.8.2
faster-whisper==0.6.0
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.2
filetype==1.2.0
fitz==0.0.1.dev2
Flask==2.3.2
Flask-SocketIO==5.3.4
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
google-api-core==1.34.0
google-auth==2.23.3
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.1.0
google-cloud-storage==2.1.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.61.0
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.0.3
groq==0.5.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.48.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.26.0
httpx-sse==0.3.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.7.0
importlib-resources==6.1.1
iniconfig==2.0.0
install==1.3.5
iopath==0.1.10
ipython==8.14.0
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kconfiglib==13.7.1
keyring==24.2.0
kiwisolver==1.4.5
langchain==0.2.1
langchain-cli==0.0.19
langchain-community==0.2.1
langchain-core==0.2.3
langchain-groq==0.1.3
langchain-openai==0.1.8
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.0
langchainhub==0.1.17
langdetect==1.0.9
langgraph==0.0.60
langserve==0.0.35
langsmith==0.1.67
layoutparser==0.3.4
libcst==1.1.0
libmagic==1.0
llama-cloud==0.0.6
llama-index-core==0.10.51
llama-index-readers-file==0.1.25
llama-parse==0.4.4
llmsherpa==0.1.4
llvmlite==0.40.1
loguru==0.7.0
looseversion==1.3.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mmh3==3.1.0
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
noiseprotocol==0.3.1
numba==0.57.1
numexpr==2.8.4
numpy==1.25.2
olefile==0.46
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.15.1
openai==1.30.1
openai-whisper==20231117
openapi-schema-pydantic==1.2.4
opencv-python==4.8.1.78
openpyxl==3.1.2
orjson==3.10.1
packaging==23.2
paho-mqtt==1.6.1
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pathspec==0.11.1
pdf2image==1.16.3
pdfkit==1.0.0
pdfminer.six==20221105
pdfplumber==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==3.2.2
pinecone-text==0.7.1
pipreqs==0.4.13
platformdirs==4.2.2
platformio==6.1.11
pluggy==1.3.0
poppler-utils==0.1.0
portalocker==2.8.2
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==3.20.3
prov==2.0.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.10
pydantic_core==2.0.1
pydot==2.0.0
pydub==0.25.1
pyelftools==0.29
Pygments==2.15.1
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypandoc==1.11
pyparsing==3.1.1
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.22.0
pyserial==3.5
pytesseract==0.3.10
pytest==7.4.4
pytest-asyncio==0.23.3
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
python-engineio==4.4.1
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
python-socketio==5.8.0
pytz==2023.3
pyxnat==1.6.2
PyYAML==6.0.1
rapidfuzz==3.4.0
rdflib==6.3.2
reedsolo==1.7.0
referencing==0.32.1
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.0
scipy==1.11.3
semantic-router==0.0.14
semantic-version==2.10.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.15
sse-starlette==1.6.1
stack-data==0.6.2
starlette==0.27.0
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.7.0
timm==0.9.8
tokenizers==0.14.1
tomlkit==0.12.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.4
traitlets==5.9.0
traits==6.3.2
transformers==4.34.1
typer==0.9.0
types-requests==2.32.0.20240602
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
unstructured==0.10.25
unstructured-inference==0.7.9
unstructured.pytesseract==0.3.12
urllib3==2.0.2
uuid6==2024.1.12
uvicorn==0.23.2
validators==0.20.0
verboselogs==1.7
virtualenv==20.26.2
voluptuous==0.14.1
wcwidth==0.2.6
weaviate-client==3.21.0
websockets==12.0
Werkzeug==2.3.4
wget==3.2
whisper==1.1.10
wkhtmltopdf==0.2
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.2
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.2
zeroconf==0.130.0
zipp==3.15.0
